backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-genai==1.14.0
influxdb-client[async]==1.48.0
redis==5.2.1
asyncpg==0.30.0
psutil==6.1.1
pydantic-settings==2.7.1
sse-starlette==2.2.1
httpx==0.28.1
numpy==2.2.2
sympy==1.13.3
scipy==1.15.2
einsteinpy==0.4.0
qutip==5.0.4
matplotlib==3.10.0
kuzu==0.11.3
pgvector==0.3.6
piper-tts==1.4.1
pyttsx3==2.99
PyYAML==6.0.2
diffusers==0.37.0
transformers
torch==2.10.0
safetensors==0.7.0
accelerate==1.13.0
pymupdf==1.25.5
python-docx==1.1.2
python-multipart==0.0.22
pytest==8.3.5
pymunk==7.2.0
tweepy==4.14.0
